NodeRef in a separate file
